查询更新profile模块代码优化
1、配置类常量名置于airflow取用，其他常量置于本项目常量模块；
2、简化部分变量名；
3、对初始化以及查询profile模块的相应os中的index的结构进行同步调整，增加search_key以及raw_data；
4、对初始化以及查询profile模块的异常处理统一profile结构；
5、增加查询更新profile模块的查询scan的scroll时间为30分钟；
6、对查询更新profile模块的redis的字典之间的数据传递由python转换为lua脚本，保证整个过程的原子性。
