+++ dags/oss_know/oss_know_dags/dags_github/dag_sync_github_profiles.py
@@ -1,7 +1,9 @@
 from datetime import datetime
 from airflow import DAG
 from airflow.operators.python import PythonOperator
-from oss_know.libs.base_dict.variable_key import OPENSEARCH_CONN_DATA, GITHUB_TOKENS
+from oss_know.libs.base_dict.variable_key import OPENSEARCH_CONN_DATA, GITHUB_TOKENS, REDIS_CLIENT_DATA, \
+    DURATION_OF_SYNC_GITHUB_PROFILES,SYNC_PROFILES_TASK_NUM
+from airflow.models import Variable
 
 # v0.0.1
 
@@ -25,18 +27,41 @@ with DAG(
 
 
     def do_sync_github_profiles():
-        from airflow.models import Variable
         from oss_know.libs.github import sync_profiles
 
         github_tokens = Variable.get(GITHUB_TOKENS, deserialize_json=True)
-        opensearch_conn_infos = Variable.get(OPENSEARCH_CONN_DATA, deserialize_json=True)
-        sync_profiles.sync_github_profiles(github_tokens, opensearch_conn_infos)
+        opensearch_conn_info = Variable.get(OPENSEARCH_CONN_DATA, deserialize_json=True)
+        redis_client_info = Variable.get(REDIS_CLIENT_DATA, deserialize_json=True)
+        duration_of_sync_github_profiles = Variable.get(DURATION_OF_SYNC_GITHUB_PROFILES, deserialize_json=True)
+        sync_profiles.sync_github_profiles(github_tokens=github_tokens, opensearch_conn_info=opensearch_conn_info,
+                                             redis_client_info=redis_client_info,
+                                             duration_of_sync_github_profiles=duration_of_sync_github_profiles)
 
 
-    op_do_sync_github_profiles = PythonOperator(
-        task_id='op_do_sync_github_profiles',
-        python_callable=do_sync_github_profiles,
+
+    # sync_profiles_task_num: 开启查询更新的线程个数
+    sync_profiles_task_num = int(Variable.get(SYNC_PROFILES_TASK_NUM, deserialize_json=True)["num"])
+    for time in range(sync_profiles_task_num):
+        op_do_sync_github_profiles = PythonOperator(
+            task_id=f'op_do_sync_github_profiles_{time + 1}',
+            python_callable=do_sync_github_profiles,
+            provide_context=True,
+        )
+
+
+    def init_storage_updated_profiles():
+        from oss_know.libs.github import sync_profiles
+
+        opensearch_conn_info = Variable.get(OPENSEARCH_CONN_DATA, deserialize_json=True)
+        redis_client_info= Variable.get(REDIS_CLIENT_DATA, deserialize_json=True)
+        sync_profiles.init_storage_pipeline(opensearch_conn_info=opensearch_conn_info,
+                                              redis_client_info=redis_client_info)
+
+
+    op_init_storage_updated_profiles = PythonOperator(
+        task_id='op_init_storage_updated_profiles',
+        python_callable=init_storage_updated_profiles,
         provide_context=True,
     )
 
-    op_scheduler_sync_github_profiles >> op_do_sync_github_profiles
+    op_scheduler_sync_github_profiles >> op_init_storage_updated_profiles >> op_do_sync_github_profiles
